requirements.txt
# GUI

streamlit
pillow>=9.0.0

# Image Processing
opencv-python>=4.7.0
numpy>=1.21.0
pdf2image>=1.16.0

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.1.0
pandas>=1.3.0
openpyxl
# OCR and AI Services
google-generativeai>=0.3.0
tiktoken>=0.5.0
pytesseract>=0.3.10
openai>=1.0.0
requests>=2.31.0

# Configuration
python-dotenv>=0.19.0
